Introduce legacy quantization mode in MLIR quantizer
The "Legacy mode" is for checking if the MLIR quantizer can output the exactly
same values with TOCO quantizer. When enabled, quantization logics are changed
to calculate scales in single-precision float, rather than default
double-precision. This makes scales exactly same for both quantizer outputs,
so that the buffers can be quantized to the same values. When double-precision
scale is used, although it might result in smaller error at the end, it's hard
to verify one-off errors in quantized values are from actual bug or slight
difference in scale.

Specifically, these values are forced to be calculated in `float`.
* Calibrated ranges. ((float)max - (float)min) / range is used.
* Bias scales. Multiplied value is downcasted to float.

Also,  quantization calls TOCO's quantization logic. TOCO quantized values with reciprocal scale (1/scale). This ensures the quantized values are the same when MLIR quantizer converted the model to correct graph and scales.

PiperOrigin-RevId: 352722201
Change-Id: I811f5b1196316cbd8d8555feaaedebe1f165ab00