[retry] Use same var key in _create_slots/get_slot in V1 optimizer
We have special handling for distributed variable in get_slot, but not
create_slot, while these keys need to match. This change modifies get_slot to use _var_key as well to avoid confusion. It is also to prepare for a upcoming refactor in dist strat code.

Note that we need to make sure the keys don't change, so existing checkpoints can still be used.

A bunch of build rules are modified to break cyclic dependencies.

PiperOrigin-RevId: 354341520
Change-Id: Ifd9786263024a11806ddde0c3bd1d36157ab8db7