Specify output type for MLIR generated kernels which have a different output type
Even if they always have the same output type, or the type is derived (e.g. complex<elem_type>), it is better to make the output type explicit.
Also make use of this possibility for kernels like Real where before we used a
trick in the template to specify the output as elem_type and the input as
complex<elem_type>.

PiperOrigin-RevId: 354942944
Change-Id: I21f9cab85189439dd8f104173aa457f64fd2d6b0