[XLA:Python] Remove unnecessary py::class_ registration for xla::LiteralSlice, and implicit conversions for literals.
These class declarations are redundant with the custom type casters for xla::LiteralSlice and xla::BorrowingLiteral.

I don't entirely recall the evolution of this code, but my guess is that I added the custom class declarations first, then discovered the need for a custom type caster. But once we have a custom type caster, we don't need the class declarations any more, and it is confusing to have both. Ultimately, the tests pass without them, so we don't need them.

PiperOrigin-RevId: 354945553
Change-Id: I11f98d2c5dbd1c3e1926c1d97cdc40944b2ce863