Consider nested ops and side effects when determining dependencies and reordering ops when Merging control flow.
A usage or successor could be nested in another op and the parent op usages need to be considered when determining whether it is safe to merge two IfRegions and whether parent op
needs to be moved after merging.  Side effects are also now considered when determining if ops need to be reordered when merging two IfRegions.

PiperOrigin-RevId: 355670174
Change-Id: Ib03544ae974df2a7f95806c9c0414ebc37968dd0