[XLA] Change xla::IsPermutation() not to check the size, and make its implementation O(n).
xla::IsPermutation() does two things: it checks a permutation is actually a permutation, and checks it has a particular size. There's no need to do the latter in xla::IsPermutation because it can be done just as easily by checking .size(), and a number of callers don't need the size check.

In addition, change the implementation to be O(n) time; there's no need to call std::is_permutation() which guarantees nothing better than O(n^2) behavior.

PiperOrigin-RevId: 356643750
Change-Id: Ib177287556b5c22266185fa813798a4ad5392054