Set default device info if device type is neither CPU nor GPU.
OpLevelCostEstimator::GetDeviceInfo() returns GFLOPs/sec and GB/sec of
the assumed machine, but currently it assumes the device type is either CPU
or GPU. We didn't expect other device types would be passed during the Grappler
optimization as transfer operations such as _Send and _Recv are added when
a graph actually runs. In some occasions, we encountered the input graph
includes _HostSend op (it is possible that the graph includes _Send or _Recv
ops; that happened in some test cases).

Device type not CPU / GPU currently causes crash; to avoid such a case, we
assume unknown device type as transfer operations over PCIe; setting
default PCIe x16 gen3 bandwidth to avoid crash.

As the GetDeviceInfo() method is virtual, one may override this method
if more precise device info is needed.

PiperOrigin-RevId: 356865551
Change-Id: Icf1f39e5da81f9c627aa01e479f437184a13eed3