Fix SparseTensorDenseMatMul test with large matrices
For a start the test wasn't being run. But it also had a problem where
the test in the complex case was trivial (the matrices) were all zeros:
it was creating `x` in `[0,1]`, then turning it into `(x+1j*x)/2` via
`_maybe_complex`, and then setting all the values with magnitude below
`0.8` to `0`. But even `(1+1j)/2` (the value with largest magntidue that
`x` can take) only has a magnitude of about `0.7`. Fix is just to create
complex numbers as `x+x*1j` instead.