Fix TFLM github CI bazel build
To keep the bazel build short, we maintain a copy of the subset of
packages that are needed for the TFLM (+ shared TfLite) bazel targets.

Eigen was updated for TF with 0effd3dc59621d79c2535e2d928d0cf41db94d3c
and we make the corresponding change in TFLM's pared down version of
workspace.bzl with this change.