Re-enable padded MatrixBatchVectorMultiply with sdot with bugfix.
Both channel_scales_ptr and row_sums_ptr are modified in the asm code
of DotprodMatrixBatchFourVectorMultiplyAccumulate. This CL marks
channel_scales_ptr as modified at the end of the asm block so the compiler
knows the register will change during execution.

More importantly, the value of channel_scales_ptr and row_sums_ptr
are now recomputed at the top of the batch loop, as their values are
only related to row and not to the batch index.

Test coverage of the padded branch was also improved in this CL.
Some of the tests that appears to test padded mode weren't really
running, because the code requires large inputs (128 * 128 or larger)
to use the padded optimization. This CL adds two large input tests.
I have verified that the new per-channel test fails without
this CL and succeeds with it.

PiperOrigin-RevId: 359561468
Change-Id: Ie499c8aef17c0ad759bbe8a8bf804ee2ccbc019e